requirements.txt
#asyncio
#random
#pathlib
#smtplib
#ssl
lxml
matplotlib
matplotlib_venn
numpy
pandas
Pillow
pyfiglet==0.8.post1
rdkit_pypi==2022.9.2
requests
seaborn==0.12.2
statannot==0.2.3
tqdm==4.59.0
protobuf==3.20.3
openai
python-dotenv
langchain
pypdf
sentence-transformers
uvicorn
ctransformers
fastapi
ipykernel
python-box
langchain-core
langchain-openai
google
pyautogen
faiss-cpu
#stripe
httpx-oauth
pyjwt
st-paywall
chromadb
mkdocs
streamlit>=1.27.0
streamlit-elements
sqlalchemy
arxiv
duckduckgo-search
wikipedia
